Add a edit_site function (#144) (#145)
* Create PVSiteEditMetadata pydantic model (#144)
Created a model with all Optinal field

* Create the edit_site function and tests (#144)
The function allow us to update site informations, if we give it a PVSiteEditMetadata instance, filled properly. If a field is set to None, it will be updated to the DB. If a field is unset, it is ignored.
Created the test_edit_site function. It test the update of a non populated field, and the update of a populated field to a None value. It also test that unfilled fields are not impacted.

* format (#144)

---------

Co-authored-by: Peter Dudfield <34686298+peterdudfield@users.noreply.github.com>